Add support for a command palette (#34)
Introduces an implementation of a command palette as a keyboard-oriented
and -- if there is a element selection -- context sensitive way of
invoking actions similar to a Code Completion / IntelliSense / Quick-Fix
functionality for code. The available commands are determined by
registered action providers that provide a set of `LabeledAction`s for a
context. A `LabeledAction` is essentially a label to be shown in the
list and an array of Sprotty `Action`s to be dispatched on invocation.

The command palette is shown on Ctrl + Space. If there is no selection
or multi-selection, it shows up in the left-upper corner of the base
div. If there is a single selection, it is shown on top of the selected
element.

The command palette builds on the previously introduced UI extension
mechanism.

Fixes #34.

Signed-off-by: Philip Langer <planger@eclipsesource.com>