Enhance context information for command palette action providers
This additional info includes the last known mouse position on the
diagram as well as the text entered into the command palette so far.
This additional info is helpful to allow for more sophisticated
and context sensitive suggestions and ordering of those suggestions.

Moreover, this change makes it easier to override the behavior for
fetching the command palette actions on typing, introduces a debounce
delay to avoid flooding the action providers with requests on each
keystroke, and finally, it adds a HTML class to the input element
while loading actions to allow visually showing that actions are
currently being loaded.

Signed-off-by: Philip Langer <planger@eclipsesource.com>