Improve loading indicator of command palette
This change simplifies showing a loading indicator while the
command palette is loading actions, e.g. with fontawesome.
It is now easy to just override loadingIndicatorClasses, which
will be assigned to a span element after the command palette
input. With CSS and fontawesome, clients can then show a
spinning icon while the actions are loading with plain css.

Signed-off-by: Philip Langer <planger@eclipsesource.com>