Avoid adding duplicate loading indicators in command palette
If for some reason the update of items fails and we never get to removing
the loading indicator, we have to avoid re-adding another one the
next time we show the command palette. Otherwise we may end up with
a remaining loading indicator that is never removed.
Also this change makes the onSelect customizable by subclasses.

Signed-off-by: Philip Langer <planger@eclipsesource.com>