Pass view args via RenderingContext
This allows omitting to pass on the specified args when `renderChildren`
is invoked. The args are always passed on via the `RenderingContext`.
If new args are specified when invoking `renderChildren`, the parent
args are still available via `args.parentArgs`. To simplify finding
a value in the args or any of its parent args, we introduce the function
`findArgValue`.

This change also introduces a type for those args `IViewArgs`
which is essentially any object, but indicates that it may have
`parentArgs`. While views may still use `object` the usage of
`IViewArgs` makes it more clear what to expect. Views may
also choose to use a more specific type.

Change-Id: I59f14d8a1bb305d63ae7b600f42cd0fe43de6b26