New Edge Router Interface `IMultipleEdgesRouter` (#247)
Summary:
New interface for edge router is proposed that extends existing one and allows using information about other elements, not only those 2 for which route is created

Details:
The current edge router interface IEdgeRouter is designed so, that each edge between two elements is routed individually. It works well in some cases, but if we want for example to avoid overlapping of other elements in a graph, then it's not enough, we need knowledge about other elements.

I propose a new edge router interface IMultipleEdgesRouter, that works not only for two elements, but also includes the context(root element) in which routing happens. Also, to optimize the routing process with multiple routers a bit, first all edges are grouped by router and then the whole group is routed at once, not each edge separately.

I work on an extension for sprotty that includes edge router based on libavoid library(wasm port of it). And proposed changes are tested in this extension, so it works. In near future I can also show its usage example.